Allow programmatic access to operations API #220
New method OperationsModule.apply(...) introduced.
